Removed types.h and split its content to macros.h and byte_buffer.h
