+++ src/client/api/restful/get_message_children.h
@@ -9,7 +9,6 @@
 #include "client/api/restful/response_error.h"
 #include "client/client_service.h"
 #include "core/models/message.h"
-#include "core/types.h"
 #include "core/utils/macros.h"
 
 /**
