+++ src/client/api/restful/get_outputs_id.h
@@ -7,7 +7,6 @@
 #include "client/api/restful/response_error.h"
 #include "client/client_service.h"
 #include "core/address.h"
-#include "core/types.h"
 #include "utarray.h"
 
 /**
