+++ src/client/api/restful/send_message.h
@@ -10,7 +10,6 @@
 #include "client/api/restful/response_error.h"
 #include "client/client_service.h"
 #include "core/models/message.h"
-#include "core/types.h"
 #include "core/utils/macros.h"
 
 /**
