+++ src/client/network/mqtt/mqtt.h
@@ -9,6 +9,7 @@
  *
  */
 
+#include <stdbool.h>
 #include <stdint.h>
 #include <stdio.h>
