+++ src/core/models/payloads/milestone.h
@@ -6,7 +6,6 @@
 
 #include <stdint.h>
 
-#include "core/types.h"
 #include "core/utils/byte_buffer.h"
 #include "crypto/iota_crypto.h"
 #include "utarray.h"
