+++ src/core/models/signing.c
@@ -1,6 +1,7 @@
 // Copyright 2022 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
+#include <stdio.h>
 #include <stdlib.h>
 
 #include "core/address.h"
